requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.27.0
python-multipart==0.0.18
langchain==0.3.7
chromadb==0.5.15
openai==1.54.0
httpx<0.28.0
python-dotenv==1.0.0
PyPDF2==3.0.1
python-docx==1.1.0
numpy>=1.24.0
huggingface_hub>=0.20.0
urllib3<2.0.0
